Remove unnecessary inclusion of #resolve-rule Remove unnecessary inclusion of #resolve-rule Return any word that matches, even if the result word is unchanged.
This fixes a logical error that occurs because Ruby's "gsub!" differs from
Clojure's "replace". If "gsub!" doesn't find a match, it returns nil. If
Clojure's "replace" doesn't find a match, it returns the original string.
Simply using "replace", there is no way to know if an match has been
made if the original and result strings are the same.

In this case, we do need to know if a match has been made, regardless of
whether or not it changes the result word.
Inflection tests and fixes for singularizing a singular.

Remove unnecessary inclusion of #resolve-rule